% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{exercise}
\alias{exercise}
\alias{exercise_data}
\alias{exercise.default}
\title{Data Exercise Structure}
\usage{
exercise(exer, ...)

\method{exercise}{default}(exer = NULL, ...)

exercise_data(exer, ...)
}
\arguments{
\item{exer}{an exercise object (default: \code{NULL})}

\item{...}{further parameters}
}
\value{
An exercise object.
}
\description{
Data structure for exercise data.
}
\examples{
exer <- exercise()           # new exercise
exer <- exercise(exer, x=3)  # add x to the exercise
}
