% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{inline}
\alias{inline}
\alias{txt_knit}
\title{Text Knitting}
\usage{
inline(txt)

txt_knit(txt)
}
\arguments{
\item{txt}{character}
}
\value{
Output.
}
\description{
Knits \code{txt} within an R code chunk.
}
\examples{
result <- inline("2 + 2")
}
