% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{pos}
\alias{pos}
\alias{dbl}
\alias{idbl}
\alias{ipos}
\alias{neg}
\alias{ineg}
\title{Interval Ranges}
\usage{
pos(pow)

neg(pow)

dbl(pow)

idbl(pow)

ipos(pow)

ineg(pow)
}
\arguments{
\item{pow}{numeric: power of ten to create intervals}
}
\value{
A \code{numeric} object.
}
\description{
Generates intervals based on powers of ten.
}
\examples{
dbl(2)
dbl(3)
pos(3)
neg(3)
}
