% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latexdef}
\alias{latexdef}
\alias{answercol}
\alias{add_answercol_def}
\title{Exam PDF with LaTeX}
\usage{
latexdef(name, body)

answercol(n)

add_answercol_def(n)
}
\arguments{
\item{name}{character: macro name}

\item{body, n}{character: macro body}
}
\value{
Nothing
}
\description{
If exams is called by \code{exams2pdf},
\itemize{
\item \code{latexdef} adds a TeX macro by \verb{\\def\\name\{body\}} and
\item \code{answercol} adds a \verb{\\def\\answercol\{n\}} to modify the number of output columns for multiple-choice answers
to the LaTeX file.
}
}
\examples{
answercol(2)
}
