% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmval.R
\name{lcmval}
\alias{lcmval}
\alias{lcm_vector}
\title{Least Common Multiple}
\usage{
lcmval(x)

lcm_vector(x)
}
\arguments{
\item{x}{integer: numbers to find the least common multiple}
}
\value{
The least common multiple.
}
\description{
Computes the least common multiple for a numeric vector \code{x}.
}
\examples{
lcmval(c(144, 160))      # = 1440
lcmval(c(144, 160, 175)) # = 50.400
}
