% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makekey.R
\name{makekey}
\alias{makekey}
\alias{make_key}
\title{Character Key Generation}
\usage{
makekey(index)

make_key(index)
}
\arguments{
\item{index}{integer: vector of integer}
}
\value{
A character.
}
\description{
Generates a character key from a vector of integers.
}
\examples{
makekey(1)
makekey(1:2)
makekey(pi) # ;)
makekey(c(5,4))
}
