% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcval.R
\name{mcval}
\alias{mcval}
\alias{compute_modes}
\alias{mcv}
\alias{mcval.default}
\alias{mcval.histogram}
\title{Most Common Value}
\usage{
mcval(x, ...)

\method{mcval}{default}(x, ...)

\method{mcval}{histogram}(x, exact = FALSE, ...)

compute_modes(x, ...)

mcv(x, ...)
}
\arguments{
\item{x}{data object}

\item{...}{further arguments}

\item{exact}{logical: either compute the exact mode or use class mids (default: \code{FALSE})}
}
\value{
A vector of modes.
}
\description{
Computes all modes (most common value).
}
\examples{
x <- sample(1:5, 15, replace=TRUE)
mcval(x)
}
