% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.R
\name{rv}
\alias{rv}
\alias{rmdFormatRV,}
\alias{lrv}
\alias{rmdFormatRV}
\title{Random Variable}
\usage{
rv(symbol, explanation)

rmdFormatRV(symbol, explanation)

lrv(symbol, explanation)
}
\arguments{
\item{symbol}{character: symbol}

\item{explanation}{character: meaning}
}
\value{
A formatted string.
}
\description{
Formats a random variable and its meaning for R Markdown.
}
\examples{
rv("X", "Waiting time in minutes until next event")
}
