% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toRMarkdown.R
\name{toRMarkdown}
\alias{toRMarkdown}
\title{toRMarkdown}
\usage{
toRMarkdown(txt)
}
\arguments{
\item{txt}{character: vector with lines of Moodle Markdown}
}
\value{
Lines with RMarkdown
}
\description{
Conversion to R Markdown.
}
\examples{
txt <- c("[image]\n",
         "Ein Paar hat 8 gute Bekannte, von denen die beiden 5 zum Essen einladen möchten.",
         "Wie viele verschiedene Reihenfolgen des Eintreffens der eingeladenen 5 Gäste gibt es?\n",
         ": 56",
         "; 120",
         "; 336",
         "; 2002",
         "; 6720",
         "; 32768",
         "; 40320",
         "; Keine Antwort ist richtig")
toRMarkdown(txt)          
}
