% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vexpar.R
\name{vexpar}
\alias{vexpar}
\title{Recursive Parsing of Expressions}
\usage{
vexpar(file, ...)
}
\arguments{
\item{file}{A single file (character) or a list of files (character vector) to be parsed.
If \code{file} is a list, the function will recursively call itself for each element in the list.}

\item{...}{Additional arguments to be passed to the underlying \code{\link[exams:expar]{exams::expar()}} function.}
}
\value{
If \code{file} is a list, the function returns a list where each element is the result of parsing a file in the list. If \code{file} is a character vector, it returns a vector of parsed expressions (or objects) obtained by applying \code{expar} to each file.
}
\description{
The \code{vexpar} function extends the functionality of \code{expar}, enabling recursive expression parsing for multiple inputs.
Unlike \code{expar}, which operates on a single file name,
\code{vexpar} can process either a single file name, a vector file names or a list of file names,
applying the parsing operation to each element automatically.
}
\details{
\itemize{
\item When a list is provided, the function uses \code{lapply} to handle each element.
\item For a character vector, the function applies \code{sapply} to parse each file using the \code{\link[exams:expar]{exams::expar()}} function.
}

This function serves as a recursive wrapper for the \code{\link[exams:expar]{exams::expar()}} function.

This documentation was created with the support of ChatGPT.
}
\examples{
\dontrun{
library("exams")
# Example usage with a single file
vexpar("file1.Rmd", instruction=NULL)
# Example usage with multiple files in a character vector
vexpar(c("file1.Rmd", "file2.Rmd"), instruction=NULL)
# Example usage with a list of files
vexpar(list("file1.Rmd", "file2.Rmd"), instruction=NULL)
}

}
