% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ts.R
\name{as_ts}
\alias{as_ts}
\title{Time Series}
\usage{
as_ts(ts)
}
\arguments{
\item{ts}{\code{ts_data} object}
}
\value{
A \code{ts} object.
}
\description{
Converts a \code{ts_data} object into a time series object (\code{ts}).
}
\examples{
# Time series from linear trend
ts <- ts_data(12, trend.coeff= c(sample(0:10, 1), sample(1+(1:10)/20, 1)))
as_ts(ts)
}
