% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catif.R
\name{catif}
\alias{catif}
\alias{condition_cat}
\title{Condition \code{cat}}
\usage{
catif(cond, ...)

condition_cat(cond, ...)
}
\arguments{
\item{cond}{logical: condition, if true then \code{cat} is called, otherwise not}

\item{...}{further parameters}
}
\value{
Invisibly \code{cond}.
}
\description{
Calls \code{cat} if \code{cond==TRUE}.
}
\examples{
catif(TRUE, "PDF")      # Should appear
catif(FALSE, "Moodle")  # Should not appear
}
